utilisation des middleware pour securiser les route

+++ app/Http/Controllers/RoleController.php
@@ -1,48 +1,52 @@
 <?php
 
-namespace App\Http\Controllers\API;
+namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use App\Models\Role;
+use App\Http\Requests\StoreRoleRequest;
+use App\Http\Requests\UpdateRoleRequest;
+use Spatie\Permission\Models\Role;
 use Spatie\Permission\Models\Permission;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 
 class RoleController extends Controller
 {
-    function __construct()
+    public function __construct()
     {
-         $this->middleware('permission:role-list|role-create|role-edit|role-delete', ['only' => ['index','store']]);
-         $this->middleware('permission:role-create', ['only' => ['create','store']]);
-         $this->middleware('permission:role-edit', ['only' => ['edit','update']]);
-         $this->middleware('permission:role-delete', ['only' => ['destroy']]);
+         $this->middleware('auth:api');
+        $this->middleware('permission:role-create|role-edit|role-delete', ['only' => ['index', 'show']]);
+        $this->middleware('permission:role-create', ['only' => ['store']]);
+        $this->middleware('permission:role-edit', ['only' => ['update']]);
+        $this->middleware('permission:role-delete', ['only' => ['destroy']]);
     }
-
 
     /**
      * Display a listing of the resource.
      */
     public function index(): JsonResponse
     {
-        $roles = Role::all();
-        return response()->json($roles, 200);
+        $roles = Role::with('permissions')->orderBy('id', 'DESC')->paginate(3);
+        return response()->json($roles);
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request): JsonResponse
+    public function store(StoreRoleRequest $request): JsonResponse
     {
-        $this->validate($request, [
-            'name' => 'required|unique:roles,name',
-            'permissions' => 'required|array',
-            'permissions.*' => 'exists:permissions,id',
-        ]);
-
-        $role = Role::create(['name' => $request->input('name')]);
-        $role->syncPermissions($request->input('permissions'));
-
-        return response()->json($role, 201); // 201 Created
+        $role = Role::create(['name' => $request->name]);
+    
+    if ($request->has('permissions') && is_array($request->permissions)) {
+        $permissions = Permission::whereIn('id', $request->permissions)->get(['name'])->toArray();
+        $role->syncPermissions($permissions);
+    }
+    
+    return response()->json([
+        'message' => 'New role is added successfully.',
+        'role' => $role
+    ], 201);
     }
 
     /**
@@ -50,25 +54,27 @@ class RoleController extends Controller
      */
     public function show(Role $role): JsonResponse
     {
-        $role->load('permissions');
-        return response()->json($role, 200);
+        return response()->json($role->load('permissions'));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Role $role): JsonResponse
+    public function update(UpdateRoleRequest $request, Role $role): JsonResponse
     {
-        $this->validate($request, [
-            'name' => 'required',
-            'permissions' => 'required|array',
-            'permissions.*' => 'exists:permissions,id',
+        if ($role->name === 'SuperAdmin') {
+            return response()->json(['message' => 'SUPER ADMIN ROLE CAN NOT BE EDITED'], 403);
+        }
+
+        $role->update($request->only('name'));
+
+        $permissions = Permission::whereIn('id', $request->permissions)->get(['name'])->toArray();
+        $role->syncPermissions($permissions);
+
+        return response()->json([
+            'message' => 'Role is updated successfully.',
+            'role' => $role
         ]);
-
-        $role->update(['name' => $request->input('name')]);
-        $role->syncPermissions($request->input('permissions'));
-
-        return response()->json($role, 200); // 200 OK
     }
 
     /**
@@ -76,7 +82,14 @@ class RoleController extends Controller
      */
     public function destroy(Role $role): JsonResponse
     {
+        if ($role->name === 'Super Admin') {
+            return response()->json(['message' => 'SUPER ADMIN ROLE CAN NOT BE DELETED'], 403);
+        }
+        if (auth()->user()->hasRole($role->name)) {
+            return response()->json(['message' => 'CAN NOT DELETE SELF ASSIGNED ROLE'], 403);
+        }
+
         $role->delete();
-        return response()->json(null, 204); // 204 No Content
+        return response()->json(['message' => 'Role is deleted successfully.']);
     }
 }
